Fixes a bug in tag_cloud() where list items were given the class '-popular' instead of 'popular' (caused by a rounding error).

+++ application/helpers/Functions.php
@@ -289,7 +289,7 @@ function tag_cloud($tags, $link = null, $maxClasses = 9)
 	foreach( $tags as $tag )
 	{
 
-		$size = ($tag["tagCount"] * $maxClasses) / $largest - 1;
+		$size = (int)(($tag["tagCount"] * $maxClasses) / $largest - 1);
 
 		$class = str_repeat('v', $size) . ($size ? '-' : '') . 'popular';
